Diseño y manejo de errores

+++ proyecto_final_moviles/lib/Producto/producto.dart
@@ -2,7 +2,7 @@ class Producto {
   int idStore;
   String nombre;
   String descripcion;
-  double precio;
+  int precio;
   String imagen;
   Producto({this.idStore,this.nombre,this.descripcion,this.precio,this.imagen});
 }
